requirements.txt
pika~=1.3.2
fastapi~=0.112.0
uvicorn~=0.30.5
mysql-connector-python~=9.0.0
pytest~=8.3.2
httpx~=0.27.0
pydantic~=2.8.2
python-dotenv~=1.0.1
strawberry-graphql[fastapi]~=0.237.3
SQLAlchemy~=1.4.53
requests~=2.32.3
